refactor: move sensor to sub-package
